requirements.txt
flask
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pillow
pysolar
pytz
requests
overpy
imagehash
geopy
pytesseract
gunicorn
